requirements.txt
# Ambiente local completo: CLI, testes (run_tests.py), transpilador, agente ADK,
# tools MCP (httpx), API FastAPI (para testes ASGI), OCR local (Pillow/pytesseract).
#
# Uso recomendado:
#   python3 -m venv .venv
#   source .venv/bin/activate   # Windows: .venv\Scripts\activate
#   pip install -U pip
#   pip install -r requirements.txt
#
# Tesseract no sistema (OCR real no host): ex. apt install tesseract-ocr tesseract-ocr-por
# (opcional; sem isso o motor OCR usa modo demo determinístico.)

google-adk>=1.20.0,<2
mcp>=1.18.0,<2
httpx>=0.27.0,<1
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
Pillow>=10.0.0,<13
pytesseract>=0.3.10,<1
jsonschema>=4.20.0,<5
python-dotenv>=1.0.0,<2
pydantic>=2.0,<3
